Add support for special "Never" type
The “Never” type is the subtype of all types.  While the bottom type
could already appear conceptually in atom types, there was previously
no robust way to handle unification of less-than-or-equal-to
constraints between Never and composite types like type applications,
tuples, and functions.  Now, Never can be a legitimate subtype of ALL
types, including composite types, and participates in type inference as
a first-class citizen.