Fix silly bug which sometimes flipped inequalities
The TypeInfer module merges inequality constraints before submitting
them to the constraint propagation system — for example, if it receives
both the constraints `x ≤ y` and `y ≤ x`, it will unify those into a
single `x = y`.   To perform this merge, it must maintain a Map
relating variable pairs to (in)equality constraints.  Variable pairs
are first re-ordered in ascending order before being used as keys into
this Map, so that `(x, y)` and `(y, x)` are not spuriously treated as
distinct keys.  However, during this reordering step, the inequality
was not flipped when the variables were flipped, so `y ≤ x` could be
entered into the map as `x ≤ y` if the `x` variable lexicographically
preceded `y`.  This has been fixed.