Add special handling for aliasing type variables.
The constraint enforcement logic in TypeInfer previously did not
account for cases where a type variable is aliased in a constraint,
such as `f = a -> a` or `t = (a, a)`.  These occur frequently in
polymorphic types, such as `id : a -> a`, `until : (a, (a -> Bool), (a
-> a)) -> a`, or `foldl1 : ((a, a) -> a), Seq<a>) -> Maybe<a>`.

Before the change, constraint enforcement steps would generate
overwrites for the two aliased variables as though they were distinct
— It would provide a new value for the left one, and a different new
value for the right one.  These values would not always match up,
because they were not generated with the knowledge that they were
related by an equality constraint, and when it came time to apply both
changes, the second would override the first and some information would
be lost.  Currently, special handling has been implemented to account
for this scenario.  It would be nice if there were some way to remove
the need for this special handling, or at least to generate some kind
of error or warning if one forgets to include it.