Guard against infinite loops caused by recursive types
If a type constraint involves a variable’s type being constructed in
terms of itself (e.g. `a = a -> a` or `a = (a, a)`), constraint
propagation will run forever.  Guarding against this is simple: just
ensure that the graph of type variables to their constituents does not
contain any cycles.