Add automatic handling of var aliasing
Previously, special handling was required when a var occurred multiple
times in a single constraint, because treating those occurrences of the
var as independent would result in it being written to twice, with the
second value overwriting the first, thereby breaking constraint
propagation.  Now, instead of handling that scenario in the constraint
enforcer code, which was a fragile and ugly hack, it’s handled
automatically in the general-purpose propagation logic.  While this is
probably slightly less performant because it involves checking for var
aliasing on every var every time a constraint is applied, it makes for
much cleaner and more robust enforcer code.

Additionally, in implementing this change, I encountered and fixed a
silly bug where “last seen versions” were never updated, and therefore
every variable was considered Changed every time a constraint was
applied.  Since the Changed/Unchanged distinction is just a performance
optimization, this bug didn’t show up until now because it manifested
only in slightly less efficient constraint propagation.