Add preliminary interaction type inference logic
Interaction type bounds are now supported in both the constraint system
and the unification system.  They are not yet exposed in any of the
external problem handling logic, so proper tests are forthcoming.
There are currently a number of non-exhaustive pattern matches caused
by the lack of interaction handling in the external problem modules.