Support invocation as standalone executable
The delta-typesolve command-line utility reads a JSON-encoded type
inference problem from stdin, and writes a JSON-encoded type inference
solution to stdout.  If the problem is syntactically well-formed but
admits no solutions (in other words, if the source code author made a
type error), the executable returns successfully with exit code 0.  If
the problem is not syntactically well-formed, the executable fails with
a nonzero exit code.