Restructure repository folder structure to hold multiple sub-projects
To accommodate additional sub-projects within the delta compiler
(parser, name resolver, code generator, etc.), all
type-inference-specific files are being moved into their own
subdirectory, and this repository’s purpose is being broadened to
include all code which executes as part of the compilation pipeline.
Originally, the intention was that every executable in the compiler
would have its own separate repository, but it is simpler to keep the
entire project within a single repository.  Not only is it simpler from
a versioning perspective for the entire compiler to exist on a single
timeline, Github-specific features like issues would be difficult to
manage for a multi-repository project.