Add `Mark` expressions for tracking source positions
To relate abstract syntax tree entities back to the original source
file (for error messages, debugging, etc), `Mark` expression nodes keep
track of the beginning and end position of expressions.  The expression
type is defined such that a valid abstract syntax tree may contain no
marks, so it is entirely at the discretion of the parsing code when and
where marks will be introduced.  The positions of syntax nodes can
often be inferred reasonably well from their children, so for now the
`Marks` inserted by the parser are fairly sparse.  More fine-grained
marking can be easily added later as needed.