Add special handling for reserved word in ParseIdent
Rather than simply categorically disallow reserved words in ParseIdent,
a more complicated system has been implemented which allows
reserved-word-checking to be enabled or disabled depending on the
context.  For example, in escapable identifiers, reserved words are
disallowed in the normal “naked” identifier form, but allowed in the
escaped form — that is actually one of the defining features of the
escaping syntax!  There is currently no escaping syntax for module
identifiers, type identifiers, or paths, but such a syntax would be
highly desirable in the future, probably using the same “`” delimiters.