Support type annotations on pattern variables
Pattern variables can now be annotated by types with the “x : A”
syntax.

Unparenthesized tuple types are not allowed as type annotations, to
avoid ambiguity with pattern-level tuples — a comma is interpreted as
ending the type and beginning another sibling pattern, not continuing
the type annotation as a tuple.

There is currently no way to annotate “ignore” patterns (“_”) with
types, but this may be added at some point if it turns out to be very
useful in some situations.