Rework Pat structure to support required type annotations
Previously, every Pat had optional type annotations — in other words,
users could write either `x` or `x : A` wherever a pattern was expected
and the pattern would be syntactically valid.  There are some
circumstances where type annotations should be required on patterns,
e.g. in the arguments of a top-level function definition.  To express
this, the Pat AST type was reworked to distinguish between Pats with
optional type annotations and Pats with required type annotations.

Instead of introducing an entirely new AST type for typed patterns, the
existing Pat AST type was parameterized by an “annotation” type, which
is `Maybe Type` in the case of optional type annotations, and `Type` in
the case of required annotations.  This minimizes code duplication.
Parsers were added for both new Pat types.

“Ignore” patterns of both pattern variants now support type annotations.