Add support for non-function defs
Syntax for defining constants which are not (necessarily) functions is
now supported in top-level declarations, protocols, and
implementations.  These constant declarations are not introduced by a
keyword; they have the same syntax as “let” statements in function
bodies.  In  top level statements and implementations, the left hand
side of a constant declaration may be a destructuring pattern.