Add support for partial dot calls and call chains
A partial dot call is a partially-applied dot-notation function,
written as a dot-notation call without a receiver.  For example, the
partial call “.foo(y)” is effectively equivalent to the function
expression “|x| {x.foo(y)}” (except that “y” is eagerly evaluated when
the function is constructed, not lazily when it is invoked).

Partial dot calls can be chained, which is equivalent to composition.
For example “.foo.bar” is effectively equivalent to “.foo >> .bar”.