Factor out logic into StripMarks module
Previously, the Syntax module exported a number of functions to strip
the source-location marks from syntax nodes.  This was provided for
testing purposes, so that parse results could be checked for equality
against expected versions which made no mention of source locations.
Because this logic was only relevant for testing, it has been factored
out into its own testing utility module, StripMarks.