Support delta-parse invocation as a standalone executable
The delta-parse subproject now includes a functioning standalone
executable version of the library.  The executable accepts as standard
input the source text of a delta module, and writes to standard out a
JSON-encoded parsed version of the module, or an error structure if it
could not be parsed.

Before the AST is written to standard out, it is stripped of
“redundant” marks — marks which enclose other marks.  These redundant
marks are generated in large quantities by certain parts of the parser,
and rather than eliminate them when they are created it is simpler just
to retroactively remove them in a single AST-wide post processing step.

In addition to being invokable as a standalone executable, the
delta-parse library now also exposes a module which can be used to
parse delta source directly into the native in-memory AST.  The
dependency on Parsec is not exposed.