Modify AST to store raw textual identifiers
Previously, identifiers were represented in the AST as
richly-structured data which precisely encoded the syntactic invariants
of the identifier types in question.  This had the advantage of
ensuring that all identifiers emitted by an AST-generating program were
syntactically valid by construction, but introduced complexity into
JSON serialization and deserialization.  Now, the rich data structures
have been retained for *constructing* identifiers, but the identifiers
themselves are raw text.  This is just as type-safe as the previous
approach, because the raw text identifier types do not expose their
constructors, and so are always produced from a rich,
invariant-encoding structure.